+++ app/src/main/java/com/app/arcabyolimpo/presentation/screens/supply/supplyUpdate/SupplyUpdateScreen.kt
@@ -44,7 +44,15 @@ import com.app.arcabyolimpo.presentation.ui.components.molecules.SelectObjectInp
 import com.app.arcabyolimpo.presentation.ui.components.molecules.StatusSelector
 import com.app.arcabyolimpo.ui.theme.Background
 import com.app.arcabyolimpo.ui.theme.White
-
+/** ---------------------------------------------------------------------------------------------- *
+ * SupplyUpdateScreen -> view where all the inputs, buttons and selects are shown with the
+ * information collected by the view model, handles different states and errors.
+ *
+ * @param viewModel: SupplyUpdateViewModel -> vm for the view in charge of placing the data
+ * @param onModifyClick: () -> Unit -> function when the user modifies the supply
+ * @param onCancel: () -> Unit -> function when the user cancels the operation
+ * @param onBackClick: () -> Unit -> function when the user clicks the back button
+ * ---------------------------------------------------------------------------------------------- */
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
 fun SupplyUpdateScreen(