+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Fixed
 - Disability's details' routes
 - Modify beneficiary (Active/inactive beneficiary)
+- Fix Decision Dialog on Delete and applied to Modify
 
 ## [0.2.0] - 02/12/2025
 ### Added
@@ -13,6 +14,7 @@
 - When added a supply it appears inmediatly in the supply list screen
 - Fix workshop video training, taking it out
 - Fix upload image in modify beneficiary
+- Fix image containers for product batches
 
 ### Updated
 - Path prefix for updating password